Move common function/vars to autoload
Functions needed by both systems and the default variables have been moved
into an autoload file, and the default variable setting has been moved to a
function and called at the beginning of both the syntax/ftplugin file.

This solves an issue where if the syntax file is loaded first (if you have
syntax on before filetype plugin on in your .vimrc), it doesn't have access to
the default variables and task state function, causing errors.